backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
msal
aiohttp
microsoft-agents-activity
microsoft-agents-copilotstudio-client